update to latest issue107-refactoring + many_to_many

+++ SIAGRO/instance_1_pgmn/graphql-server/resolvers/individual.js
@@ -14,6 +14,8 @@ const os = require('os');
 const resolvers = require(path.join(__dirname, 'index.js'));
 const models = require(path.join(__dirname, '..', 'models_index.js'));
 
+
+
 const associationArgsDef = {
     'addAccession': 'accession',
     'addMeasurements': 'measurement'
@@ -31,35 +33,36 @@ const associationArgsDef = {
 individual.prototype.accession = async function({
     search
 }, context) {
-    try {
-        if (search === undefined) {
-            return resolvers.readOneAccession({
-                [models.accession.idAttribute()]: this.accession_id
-            }, context)
-        } else {
-            //build new search filter
-            let nsearch = helper.addSearchField({
-                "search": search,
-                "field": models.accession.idAttribute(),
-                "value": {
-                    "value": this.accession_id
-                },
-                "operator": "eq"
-            });
-            let found = await resolvers.accessions({
-                search: nsearch
-            }, context);
-            if (found) {
-                return found[0]
+    if (helper.isNotUndefinedAndNotNull(this.accession_id)) {
+        try {
+            if (search === undefined) {
+                return resolvers.readOneAccession({
+                    [models.accession.idAttribute()]: this.accession_id
+                }, context)
+            } else {
+                //build new search filter
+                let nsearch = helper.addSearchField({
+                    "search": search,
+                    "field": models.accession.idAttribute(),
+                    "value": {
+                        "value": this.accession_id
+                    },
+                    "operator": "eq"
+                });
+                let found = await resolvers.accessions({
+                    search: nsearch
+                }, context);
+                if (found) {
+                    return found[0]
+                }
+                return found;
             }
-            return found;
-        }
-    } catch (error) {
-        console.error(error);
-        handleError(error);
-    };
+        } catch (error) {
+            console.error(error);
+            handleError(error);
+        };
+    }
 }
-
 
 /**
  * individual.prototype.measurementsFilter - Check user authorization and return certain number, specified in pagination argument, of records
@@ -130,7 +133,6 @@ individual.prototype.countFilteredMeasurements = function({
     };
 }
 
-
 /**
  * individual.prototype.measurementsConnection - Check user authorization and return certain number, specified in pagination argument, of records
  * associated with the current instance, this records should also
@@ -180,13 +182,17 @@ individual.prototype.measurementsConnection = function({
 individual.prototype.handleAssociations = async function(input, context) {
     try {
         let promises = [];
-
         if (helper.isNonEmptyArray(input.addMeasurements)) {
             promises.push(this.addMeasurements(input, context));
         }
-
         if (helper.isNotUndefinedAndNotNull(input.addAccession)) {
             promises.push(this.addAccession(input, context));
+        }
+        if (helper.isNonEmptyArray(input.removeMeasurements)) {
+            promises.push(this.removeMeasurements(input, context));
+        }
+        if (helper.isNotUndefinedAndNotNull(input.removeAccession)) {
+            promises.push(this.removeAccession(input, context));
         }
 
         await Promise.all(promises);
@@ -195,58 +201,56 @@ individual.prototype.handleAssociations = async function(input, context) {
     }
 }
 
-
-
 /**
- * addMeasurements - field Mutation for to_many associations to add 
+ * addMeasurements - field Mutation for to_many associationsArguments to add 
  *
  * @param {object} input   Info of input Ids to add  the association
  */
 individual.prototype.addMeasurements = async function(input) {
     let results = [];
     input.addMeasurements.forEach(associatedRecordId => {
-        results.push(models.measurement._addIndividual(associatedRecordId, input.name));
+        results.push(models.measurement._addIndividual(associatedRecordId, this.getIdValue()));
     })
     await Promise.all(results);
 }
 
-
 /**
- * addAccession - field Mutation for to_one associations to add 
+ * addAccession - field Mutation for to_one associationsArguments to add 
  *
  * @param {object} input   Info of input Ids to add  the association
  */
 individual.prototype.addAccession = async function(input) {
-    await individual._addAccession(input.name, input.addAccession);
+    await individual._addAccession(this.getIdValue(), input.addAccession);
     this.accession_id = input.addAccession;
 }
 
 
 
-
 /**
- * removeMeasurements - field Mutation for to_many associations to remove 
+ * removeMeasurements - field Mutation for to_many associationsArguments to remove 
  *
  * @param {object} input   Info of input Ids to remove  the association
  */
 individual.prototype.removeMeasurements = async function(input) {
     let results = [];
     input.removeMeasurements.forEach(associatedRecordId => {
-        results.push(models.measurement._removeIndividual(associatedRecordId, input.name));
+        results.push(models.measurement._removeIndividual(associatedRecordId, this.getIdValue()));
     })
     await Promise.all(results);
 }
 
-
 /**
- * removeAccession - field Mutation for to_one associations to remove 
+ * removeAccession - field Mutation for to_one associationsArguments to remove 
  *
  * @param {object} input   Info of input Ids to remove  the association
  */
 individual.prototype.removeAccession = async function(input) {
-    await individual._removeAccession(input.name, input.removeAccession);
-    this.accession_id = input.removeAccession;
+    if (input.removeAccession === this.accession_id) {
+        await individual._removeAccession(this.getIdValue(), input.removeAccession);
+        this.accession_id = null;
+    }
 }
+
 
 /**
  * errorMessageForRecordsLimit(query) - returns error message in case the record limit is exceeded.
@@ -272,7 +276,7 @@ async function checkCount(search, context, query) {
 
 /**
  * checkCountForOne(context) - Make sure that the record limit is not exhausted before requesting a single record
- * 
+ *
  * @param {object} context Provided to every resolver holds contextual information like the resquest query and user info.
  */
 function checkCountForOne(context) {
@@ -295,11 +299,50 @@ function checkCountAgainAndAdaptLimit(context, numberOfFoundItems, query) {
     context.recordsLimit -= numberOfFoundItems;
 }
 
+/**
+ * countAllAssociatedRecords - Count records associated with another given record
+ *
+ * @param  {ID} id      Id of the record which the associations will be counted
+ * @param  {objec} context Default context by resolver
+ * @return {Int}         Number of associated records
+ */
+async function countAllAssociatedRecords(id, context) {
 
+    let individual = await resolvers.readOneIndividual({
+        name: id
+    }, context);
+    //check that record actually exists
+    if (individual === null) throw new Error(`Record with ID = ${id} does not exist`);
+    let promises_to_many = [];
+    let promises_to_one = [];
 
+    promises_to_many.push(individual.countFilteredMeasurements({}, context));
+    promises_to_one.push(individual.accession({}, context));
+
+    let result_to_many = await Promise.all(promises_to_many);
+    let result_to_one = await Promise.all(promises_to_one);
+
+    let get_to_many_associated = result_to_many.reduce((accumulator, current_val) => accumulator + current_val, 0);
+    let get_to_one_associated = result_to_one.filter((r, index) => r !== null).length;
+
+    return get_to_one_associated + get_to_many_associated;
+}
+
+/**
+ * validForDeletion - Checks wether a record is allowed to be deleted
+ *
+ * @param  {ID} id      Id of record to check if it can be deleted
+ * @param  {object} context Default context by resolver
+ * @return {boolean}         True if it is allowed to be deleted and false otherwise
+ */
+async function validForDeletion(id, context) {
+    if (await countAllAssociatedRecords(id, context) > 0) {
+        throw new Error(`Accession with accession_id ${id} has associated records and is NOT valid for deletion. Please clean up before you delete.`);
+    }
+    return true;
+}
 
 module.exports = {
-
     /**
      * individuals - Check user authorization and return certain number, specified in pagination argument, of records that
      * holds the condition of search argument, all of them sorted as specified by the order argument.
@@ -360,7 +403,6 @@ module.exports = {
         })
     },
 
-
     /**
      * readOneIndividual - Check user authorization and return one record with the specified name in the name argument.
      *
@@ -378,6 +420,48 @@ module.exports = {
                 checkCountForOne(context);
                 context.recordsLimit = context.recordsLimit - 1;
                 return resultRecords;
+            } else {
+                throw new Error("You don't have authorization to perform this action");
+            }
+        }).catch(error => {
+            console.error(error);
+            handleError(error);
+        })
+    },
+
+    /**
+     * countIndividuals - Counts number of records that holds the conditions specified in the search argument
+     *
+     * @param  {object} {search} Search argument for filtering records
+     * @param  {object} context  Provided to every resolver holds contextual information like the resquest query and user info.
+     * @return {number}          Number of records that holds the conditions specified in the search argument
+     */
+    countIndividuals: function({
+        search
+    }, context) {
+        return checkAuthorization(context, 'Individual', 'read').then(authorization => {
+            if (authorization === true) {
+                return individual.countRecords(search);
+            } else {
+                throw new Error("You don't have authorization to perform this action");
+            }
+        }).catch(error => {
+            console.error(error);
+            handleError(error);
+        })
+    },
+
+    /**
+     * vueTableIndividual - Return table of records as needed for displaying a vuejs table
+     *
+     * @param  {string} _       First parameter is not used
+     * @param  {object} context Provided to every resolver holds contextual information like the resquest query and user info.
+     * @return {object}         Records with format as needed for displaying a vuejs table
+     */
+    vueTableIndividual: function(_, context) {
+        return checkAuthorization(context, 'Individual', 'read').then(authorization => {
+            if (authorization === true) {
+                return helper.vueTable(context.request, individual, ["id", "name", "origin", "description", "accession_id"]);
             } else {
                 throw new Error("You don't have authorization to perform this action");
             }
@@ -445,9 +529,11 @@ module.exports = {
     deleteIndividual: function({
         name
     }, context) {
-        return checkAuthorization(context, 'Individual', 'delete').then(authorization => {
+        return checkAuthorization(context, 'Individual', 'delete').then(async authorization => {
             if (authorization === true) {
-                return individual.deleteOne(name);
+                if (await individual.validForDeletion(name, context)) {
+                    return individual.deleteOne(name);
+                }
             } else {
                 throw new Error("You don't have authorization to perform this action");
             }
@@ -484,48 +570,6 @@ module.exports = {
             console.error(error);
             handleError(error);
         }
-    },
-
-    /**
-     * countIndividuals - Counts number of records that holds the conditions specified in the search argument
-     *
-     * @param  {object} {search} Search argument for filtering records
-     * @param  {object} context  Provided to every resolver holds contextual information like the resquest query and user info.
-     * @return {number}          Number of records that holds the conditions specified in the search argument
-     */
-    countIndividuals: function({
-        search
-    }, context) {
-        return checkAuthorization(context, 'Individual', 'read').then(authorization => {
-            if (authorization === true) {
-                return individual.countRecords(search);
-            } else {
-                throw new Error("You don't have authorization to perform this action");
-            }
-        }).catch(error => {
-            console.error(error);
-            handleError(error);
-        })
-    },
-
-    /**
-     * vueTableIndividual - Return table of records as needed for displaying a vuejs table
-     *
-     * @param  {string} _       First parameter is not used
-     * @param  {object} context Provided to every resolver holds contextual information like the resquest query and user info.
-     * @return {object}         Records with format as needed for displaying a vuejs table
-     */
-    vueTableIndividual: function(_, context) {
-        return checkAuthorization(context, 'Individual', 'read').then(authorization => {
-            if (authorization === true) {
-                return helper.vueTable(context.request, individual, ["id", "name", "origin", "description", "accession_id"]);
-            } else {
-                throw new Error("You don't have authorization to perform this action");
-            }
-        }).catch(error => {
-            console.error(error);
-            handleError(error);
-        })
     },
 
     /**
