+++ SIAGRO/instance_1_pgmn/graphql-server/resolvers/measurement.js
@@ -14,6 +14,8 @@ const os = require('os');
 const resolvers = require(path.join(__dirname, 'index.js'));
 const models = require(path.join(__dirname, '..', 'models_index.js'));
 
+
+
 const associationArgsDef = {
     'addIndividual': 'individual',
     'addAccession': 'accession'
@@ -31,33 +33,35 @@ const associationArgsDef = {
 measurement.prototype.individual = async function({
     search
 }, context) {
-    try {
-        if (search === undefined) {
-            return resolvers.readOneIndividual({
-                [models.individual.idAttribute()]: this.individual_id
-            }, context)
-        } else {
-            //build new search filter
-            let nsearch = helper.addSearchField({
-                "search": search,
-                "field": models.individual.idAttribute(),
-                "value": {
-                    "value": this.individual_id
-                },
-                "operator": "eq"
-            });
-            let found = await resolvers.individuals({
-                search: nsearch
-            }, context);
-            if (found) {
-                return found[0]
+    if (helper.isNotUndefinedAndNotNull(this.individual_id)) {
+        try {
+            if (search === undefined) {
+                return resolvers.readOneIndividual({
+                    [models.individual.idAttribute()]: this.individual_id
+                }, context)
+            } else {
+                //build new search filter
+                let nsearch = helper.addSearchField({
+                    "search": search,
+                    "field": models.individual.idAttribute(),
+                    "value": {
+                        "value": this.individual_id
+                    },
+                    "operator": "eq"
+                });
+                let found = await resolvers.individuals({
+                    search: nsearch
+                }, context);
+                if (found) {
+                    return found[0]
+                }
+                return found;
             }
-            return found;
-        }
-    } catch (error) {
-        console.error(error);
-        handleError(error);
-    };
+        } catch (error) {
+            console.error(error);
+            handleError(error);
+        };
+    }
 }
 /**
  * measurement.prototype.accession - Return associated record
@@ -69,33 +73,35 @@ measurement.prototype.individual = async function({
 measurement.prototype.accession = async function({
     search
 }, context) {
-    try {
-        if (search === undefined) {
-            return resolvers.readOneAccession({
-                [models.accession.idAttribute()]: this.accession_id
-            }, context)
-        } else {
-            //build new search filter
-            let nsearch = helper.addSearchField({
-                "search": search,
-                "field": models.accession.idAttribute(),
-                "value": {
-                    "value": this.accession_id
-                },
-                "operator": "eq"
-            });
-            let found = await resolvers.accessions({
-                search: nsearch
-            }, context);
-            if (found) {
-                return found[0]
+    if (helper.isNotUndefinedAndNotNull(this.accession_id)) {
+        try {
+            if (search === undefined) {
+                return resolvers.readOneAccession({
+                    [models.accession.idAttribute()]: this.accession_id
+                }, context)
+            } else {
+                //build new search filter
+                let nsearch = helper.addSearchField({
+                    "search": search,
+                    "field": models.accession.idAttribute(),
+                    "value": {
+                        "value": this.accession_id
+                    },
+                    "operator": "eq"
+                });
+                let found = await resolvers.accessions({
+                    search: nsearch
+                }, context);
+                if (found) {
+                    return found[0]
+                }
+                return found;
             }
-            return found;
-        }
-    } catch (error) {
-        console.error(error);
-        handleError(error);
-    };
+        } catch (error) {
+            console.error(error);
+            handleError(error);
+        };
+    }
 }
 
 
@@ -110,12 +116,18 @@ measurement.prototype.handleAssociations = async function(input, context) {
     try {
         let promises = [];
 
-
         if (helper.isNotUndefinedAndNotNull(input.addIndividual)) {
             promises.push(this.addIndividual(input, context));
         }
         if (helper.isNotUndefinedAndNotNull(input.addAccession)) {
             promises.push(this.addAccession(input, context));
+        }
+
+        if (helper.isNotUndefinedAndNotNull(input.removeIndividual)) {
+            promises.push(this.removeIndividual(input, context));
+        }
+        if (helper.isNotUndefinedAndNotNull(input.removeAccession)) {
+            promises.push(this.removeAccession(input, context));
         }
 
         await Promise.all(promises);
@@ -124,50 +136,52 @@ measurement.prototype.handleAssociations = async function(input, context) {
     }
 }
 
-
-
-
 /**
- * addIndividual - field Mutation for to_one associations to add 
+ * addIndividual - field Mutation for to_one associationsArguments to add 
  *
  * @param {object} input   Info of input Ids to add  the association
  */
 measurement.prototype.addIndividual = async function(input) {
-    await measurement._addIndividual(input.measurement_id, input.addIndividual);
+    await measurement._addIndividual(this.getIdValue(), input.addIndividual);
     this.individual_id = input.addIndividual;
 }
+
 /**
- * addAccession - field Mutation for to_one associations to add 
+ * addAccession - field Mutation for to_one associationsArguments to add 
  *
  * @param {object} input   Info of input Ids to add  the association
  */
 measurement.prototype.addAccession = async function(input) {
-    await measurement._addAccession(input.measurement_id, input.addAccession);
+    await measurement._addAccession(this.getIdValue(), input.addAccession);
     this.accession_id = input.addAccession;
 }
 
 
 
-
-
 /**
- * removeIndividual - field Mutation for to_one associations to remove 
+ * removeIndividual - field Mutation for to_one associationsArguments to remove 
  *
  * @param {object} input   Info of input Ids to remove  the association
  */
 measurement.prototype.removeIndividual = async function(input) {
-    await measurement._removeIndividual(input.measurement_id, input.removeIndividual);
-    this.individual_id = input.removeIndividual;
+    if (input.removeIndividual === this.individual_id) {
+        await measurement._removeIndividual(this.getIdValue(), input.removeIndividual);
+        this.individual_id = null;
+    }
 }
+
 /**
- * removeAccession - field Mutation for to_one associations to remove 
+ * removeAccession - field Mutation for to_one associationsArguments to remove 
  *
  * @param {object} input   Info of input Ids to remove  the association
  */
 measurement.prototype.removeAccession = async function(input) {
-    await measurement._removeAccession(input.measurement_id, input.removeAccession);
-    this.accession_id = input.removeAccession;
+    if (input.removeAccession === this.accession_id) {
+        await measurement._removeAccession(this.getIdValue(), input.removeAccession);
+        this.accession_id = null;
+    }
 }
+
 
 /**
  * errorMessageForRecordsLimit(query) - returns error message in case the record limit is exceeded.
@@ -193,7 +207,7 @@ async function checkCount(search, context, query) {
 
 /**
  * checkCountForOne(context) - Make sure that the record limit is not exhausted before requesting a single record
- * 
+ *
  * @param {object} context Provided to every resolver holds contextual information like the resquest query and user info.
  */
 function checkCountForOne(context) {
@@ -216,11 +230,50 @@ function checkCountAgainAndAdaptLimit(context, numberOfFoundItems, query) {
     context.recordsLimit -= numberOfFoundItems;
 }
 
+/**
+ * countAllAssociatedRecords - Count records associated with another given record
+ *
+ * @param  {ID} id      Id of the record which the associations will be counted
+ * @param  {objec} context Default context by resolver
+ * @return {Int}         Number of associated records
+ */
+async function countAllAssociatedRecords(id, context) {
 
+    let measurement = await resolvers.readOneMeasurement({
+        measurement_id: id
+    }, context);
+    //check that record actually exists
+    if (measurement === null) throw new Error(`Record with ID = ${id} does not exist`);
+    let promises_to_many = [];
+    let promises_to_one = [];
 
+    promises_to_one.push(measurement.individual({}, context));
+    promises_to_one.push(measurement.accession({}, context));
+
+    let result_to_many = await Promise.all(promises_to_many);
+    let result_to_one = await Promise.all(promises_to_one);
+
+    let get_to_many_associated = result_to_many.reduce((accumulator, current_val) => accumulator + current_val, 0);
+    let get_to_one_associated = result_to_one.filter((r, index) => r !== null).length;
+
+    return get_to_one_associated + get_to_many_associated;
+}
+
+/**
+ * validForDeletion - Checks wether a record is allowed to be deleted
+ *
+ * @param  {ID} id      Id of record to check if it can be deleted
+ * @param  {object} context Default context by resolver
+ * @return {boolean}         True if it is allowed to be deleted and false otherwise
+ */
+async function validForDeletion(id, context) {
+    if (await countAllAssociatedRecords(id, context) > 0) {
+        throw new Error(`Accession with accession_id ${id} has associated records and is NOT valid for deletion. Please clean up before you delete.`);
+    }
+    return true;
+}
 
 module.exports = {
-
     /**
      * measurements - Check user authorization and return certain number, specified in pagination argument, of records that
      * holds the condition of search argument, all of them sorted as specified by the order argument.
@@ -281,7 +334,6 @@ module.exports = {
         })
     },
 
-
     /**
      * readOneMeasurement - Check user authorization and return one record with the specified measurement_id in the measurement_id argument.
      *
@@ -299,6 +351,48 @@ module.exports = {
                 checkCountForOne(context);
                 context.recordsLimit = context.recordsLimit - 1;
                 return resultRecords;
+            } else {
+                throw new Error("You don't have authorization to perform this action");
+            }
+        }).catch(error => {
+            console.error(error);
+            handleError(error);
+        })
+    },
+
+    /**
+     * countMeasurements - Counts number of records that holds the conditions specified in the search argument
+     *
+     * @param  {object} {search} Search argument for filtering records
+     * @param  {object} context  Provided to every resolver holds contextual information like the resquest query and user info.
+     * @return {number}          Number of records that holds the conditions specified in the search argument
+     */
+    countMeasurements: function({
+        search
+    }, context) {
+        return checkAuthorization(context, 'Measurement', 'read').then(authorization => {
+            if (authorization === true) {
+                return measurement.countRecords(search);
+            } else {
+                throw new Error("You don't have authorization to perform this action");
+            }
+        }).catch(error => {
+            console.error(error);
+            handleError(error);
+        })
+    },
+
+    /**
+     * vueTableMeasurement - Return table of records as needed for displaying a vuejs table
+     *
+     * @param  {string} _       First parameter is not used
+     * @param  {object} context Provided to every resolver holds contextual information like the resquest query and user info.
+     * @return {object}         Records with format as needed for displaying a vuejs table
+     */
+    vueTableMeasurement: function(_, context) {
+        return checkAuthorization(context, 'Measurement', 'read').then(authorization => {
+            if (authorization === true) {
+                return helper.vueTable(context.request, measurement, ["id", "measurement_id", "name", "method", "reference", "reference_link", "unit", "short_name", "comments", "individual_id", "accession_id"]);
             } else {
                 throw new Error("You don't have authorization to perform this action");
             }
@@ -366,9 +460,11 @@ module.exports = {
     deleteMeasurement: function({
         measurement_id
     }, context) {
-        return checkAuthorization(context, 'Measurement', 'delete').then(authorization => {
+        return checkAuthorization(context, 'Measurement', 'delete').then(async authorization => {
             if (authorization === true) {
-                return measurement.deleteOne(measurement_id);
+                if (await measurement.validForDeletion(measurement_id, context)) {
+                    return measurement.deleteOne(measurement_id);
+                }
             } else {
                 throw new Error("You don't have authorization to perform this action");
             }
@@ -405,48 +501,6 @@ module.exports = {
             console.error(error);
             handleError(error);
         }
-    },
-
-    /**
-     * countMeasurements - Counts number of records that holds the conditions specified in the search argument
-     *
-     * @param  {object} {search} Search argument for filtering records
-     * @param  {object} context  Provided to every resolver holds contextual information like the resquest query and user info.
-     * @return {number}          Number of records that holds the conditions specified in the search argument
-     */
-    countMeasurements: function({
-        search
-    }, context) {
-        return checkAuthorization(context, 'Measurement', 'read').then(authorization => {
-            if (authorization === true) {
-                return measurement.countRecords(search);
-            } else {
-                throw new Error("You don't have authorization to perform this action");
-            }
-        }).catch(error => {
-            console.error(error);
-            handleError(error);
-        })
-    },
-
-    /**
-     * vueTableMeasurement - Return table of records as needed for displaying a vuejs table
-     *
-     * @param  {string} _       First parameter is not used
-     * @param  {object} context Provided to every resolver holds contextual information like the resquest query and user info.
-     * @return {object}         Records with format as needed for displaying a vuejs table
-     */
-    vueTableMeasurement: function(_, context) {
-        return checkAuthorization(context, 'Measurement', 'read').then(authorization => {
-            if (authorization === true) {
-                return helper.vueTable(context.request, measurement, ["id", "measurement_id", "name", "method", "reference", "reference_link", "unit", "short_name", "comments", "individual_id", "accession_id"]);
-            } else {
-                throw new Error("You don't have authorization to perform this action");
-            }
-        }).catch(error => {
-            console.error(error);
-            handleError(error);
-        })
     },
 
     /**
